+++ app/src/main/java/cn/innovativest/ath/widget/TradeBuyDialog.java
@@ -107,7 +107,7 @@ public class TradeBuyDialog extends Dialog implements View.OnClickListener {
             String str[] = tradeItem.getQuota().split(",");
             if (str.length == 2) {
                 edtSDSD.setText("");
-                edtSDSD.setHint(AppUtils.floatToStringByTruncate(Float.parseFloat(str[0]) / Float.parseFloat(tradeItem.getCny()), 6) + " - " + AppUtils.floatToStringByTruncate(Float.parseFloat(str[1]) / Float.parseFloat(tradeItem.getCny()), 6));
+                edtSDSD.setHint(AppUtils.floatToStringByTruncate(Double.parseDouble(str[0]) / Double.parseDouble(tradeItem.getCny()), 6) + " - " + AppUtils.floatToStringByTruncate(Double.parseDouble(str[1]) / Double.parseDouble(tradeItem.getCny()), 6));
             }
         }
         dgAmount.setText(String.format("%.2f", Float.parseFloat(tradeItem.getCny())) + " CNY");
@@ -195,7 +195,7 @@ public class TradeBuyDialog extends Dialog implements View.OnClickListener {
                     if (!CUtils.isEmpty(tradeItem.getQuota()) && tradeItem.getQuota().contains(",")) {
                         String str[] = tradeItem.getQuota().split(",");
                         if (str.length == 2) {
-                            if (Float.parseFloat(edtSDSD.getText().toString()) > Float.parseFloat(str[1]) / Float.parseFloat(tradeItem.getCny()) || Float.parseFloat(edtSDSD.getText().toString()) < Float.parseFloat(str[0]) / Float.parseFloat(tradeItem.getCny())) {
+                            if (Double.parseDouble(edtSDSD.getText().toString()) > Double.parseDouble(str[1]) / Double.parseDouble(tradeItem.getCny()) || Double.parseDouble(edtSDSD.getText().toString()) < Double.parseDouble(str[0]) / Double.parseDouble(tradeItem.getCny())) {
                                 App.toast(getContext(), "输入数量不能小于最小限额或大于最大限额");
                                 return;
                             }
@@ -217,8 +217,8 @@ public class TradeBuyDialog extends Dialog implements View.OnClickListener {
                     if (str.length == 2) {
 //                        edtSDSD.setText(String.format("%.6f", Float.parseFloat(str[1]) / Float.parseFloat(tradeItem.getCny())));
 //                        edtSDSD.setSelection(String.format("%.6f", Float.parseFloat(str[1]) / Float.parseFloat(tradeItem.getCny())).length());
-                        edtSDSD.setText(AppUtils.floatToStringByTruncate(Float.parseFloat(str[1]) / Float.parseFloat(tradeItem.getCny()), 6));
-                        edtSDSD.setSelection(AppUtils.floatToStringByTruncate(Float.parseFloat(str[1]) / Float.parseFloat(tradeItem.getCny()), 6).length());
+                        edtSDSD.setText(AppUtils.floatToStringByTruncate(Double.parseDouble(str[1]) / Double.parseDouble(tradeItem.getCny()), 6));
+                        edtSDSD.setSelection(AppUtils.floatToStringByTruncate(Double.parseDouble(str[1]) / Double.parseDouble(tradeItem.getCny()), 6).length());
                     }
                 }
                 break;
